# openapi.yaml
openapi: 3.0.3
info:
  title: FinTrust Secure Platform API
  description: |
    Hardened Open Banking–style reference API. Refresh tokens are issued as HttpOnly cookies
    (`Path=/auth`) in addition to short-lived access tokens returned in JSON.
  version: 1.0.0
servers:
  - url: https://api.fintrust.local
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
paths:
  /auth/register:
    post:
      summary: Register a standard user (role=user)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  minLength: 12
      responses:
        "201":
          description: Created
        "400":
          description: Validation or password policy failure
        "409":
          description: Email already registered
  /auth/login:
    post:
      summary: Login; refresh token is set as HttpOnly cookie
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
      responses:
        "200":
          description: Access token issued; refresh cookie set
        "401":
          description: Invalid credentials
        "423":
          description: Account locked
  /auth/refresh:
    post:
      summary: Rotate refresh token (HttpOnly cookie only) and issue new access token
      responses:
        "200":
          description: Tokens rotated
        "401":
          description: Invalid or revoked refresh
...